+++ src/graph.c
@@ -8,38 +8,38 @@
 // This edge struct acts as a linked list node
 typedef struct edge
 {
-    uint64_t destination; /* destination vertex of edge a.k.a an adjascent vertex */
+    uint32_t destination; /* destination vertex of edge a.k.a an adjascent vertex */
     struct edge *next;    /* next edge in list... I have to use the struct keyboard because this is recursive */
 } edge;
 
 typedef struct graph
 {
-    uint64_t number_vertices; /* number of vertices in graph */
+    uint32_t number_vertices; /* number of vertices in graph */
     uint64_t number_edges;    /* number of edges in graph */
     bool is_directed;         /* Is the graph directed */
-    uint64_t *IA;             /* Cumulative count of non zero values excluding the row at this index */
-    uint64_t *JA;             /* Actual Column Indices of non zero values. Use IA first... */
+    uint32_t *IA;             /* Cumulative count of non zero values excluding the row at this index */
+    uint32_t *JA;             /* Actual Column Indices of non zero values. Use IA first... */
 
 } graph;
 
 void initialize_graph(graph *g, bool directed);
 void read_graph(graph *g, bool directed);
-void insert_edge(graph *g, edge *edges[], uint64_t x, uint64_t y, bool directed);
+void insert_edge(graph *g, edge *edges[], uint32_t x, uint32_t y, bool directed);
 void print_graph(graph *g);
 void build_csr(graph *g, edge *edges[]);
 
-uint64_t getDegree(graph *g, uint64_t vertex);
-void getNeighbors(graph *g, uint64_t source, uint64_t results[]);
+uint64_t getDegree(graph *g, uint32_t vertex);
+void getNeighbors(graph *g, uint32_t source, uint32_t results[]);
 
-uint64_t getDegree(graph *g, uint64_t vertex)
+uint64_t getDegree(graph *g, uint32_t vertex)
 {
     return g->IA[vertex + 1] - g->IA[vertex];
 }
 
 // Takes a pointer to an array and writes out all neighbors from CSR
-void getNeighbors(graph *g, uint64_t source, uint64_t results[])
+void getNeighbors(graph *g, uint32_t source, uint32_t results[])
 {
-    for (uint64_t idx = 0; idx + g->IA[source] < g->IA[source + 1]; idx++)
+    for (uint32_t idx = 0; idx + g->IA[source] < g->IA[source + 1]; idx++)
     {
         results[idx] = g->JA[idx + g->IA[source]];
     }
@@ -57,13 +57,13 @@ void read_graph(graph *g, bool is_directed)
 {
     edge *edges[MAXV + 1];        /* Array of Pointers used to buffer adjasency list as itermediate representation before building CSR*/
     uint64_t edge_count;          /* number of edges */
-    uint64_t source, destination; /* vertices in edge (source, destination) */
+    uint32_t source, destination; /* vertices in edge (source, destination) */
 
-    scanf("%lu %lu", &(g->number_vertices), &edge_count);
+    scanf("%u %lu", &(g->number_vertices), &edge_count);
 
     for (uint64_t i = 1; i <= edge_count; i++)
     {
-        scanf("%lu %lu", &source, &destination);
+        scanf("%u %u", &source, &destination);
         insert_edge(g, edges, source, destination, is_directed);
     }
     build_csr(g, edges);
@@ -89,16 +89,16 @@ void read_graph(graph *g, bool is_directed)
 
 void build_csr(graph *g, edge *edges[])
 {
-    g->IA = malloc((g->number_edges + 2) * sizeof(uint64_t)); /* I might be off by one here */
+    g->IA = malloc((g->number_edges + 2) * sizeof(uint32_t)); /* I might be off by one here */
 
     // Mark the 0 positions as Evil to force a crash to ensure I'm indexing starting at 1 to be consistent
     g->IA[0] = 666;
 
     g->IA[1] = 0;
     // I'm just multiplying by 2 here because the number_edges is only incremented one for the opposing edges on a non-DiGraph. I should really only store non-directed edges once.
-    g->JA = malloc((g->number_edges * 2) * sizeof(uint64_t)); // "Column Indices"
+    g->JA = malloc((g->number_edges * 2) * sizeof(uint32_t)); // "Column Indices"
     uint64_t edge_idx = 0;
-    for (uint64_t current_vertex = 1; current_vertex < g->number_vertices; current_vertex++)
+    for (uint32_t current_vertex = 1; current_vertex < g->number_vertices; current_vertex++)
     {
         for (edge *currentEdge = edges[current_vertex]; currentEdge->next != NULL; currentEdge = currentEdge->next)
         {
@@ -110,7 +110,7 @@ void build_csr(graph *g, edge *edges[])
 }
 
 // Inserts an edge from source to destination in the adjacency list of graph g. If the edge is not directed, it adds source -> destination and destination -> but only increments th edge count once.
-void insert_edge(graph *g, edge *edges[], uint64_t source, uint64_t destination, bool is_directed)
+void insert_edge(graph *g, edge *edges[], uint32_t source, uint32_t destination, bool is_directed)
 {
     edge *new_edge = malloc(sizeof(edge)); /* temporary pointer */
     new_edge->destination = destination;
@@ -131,12 +131,12 @@ void insert_edge(graph *g, edge *edges[], uint64_t source, uint64_t destination,
 
 void print_graph(graph *g)
 {
-    for (uint64_t current_vertex = 1; current_vertex <= g->number_vertices; current_vertex++)
+    for (uint32_t current_vertex = 1; current_vertex <= g->number_vertices; current_vertex++)
     {
-        printf("%lu: ", current_vertex);
+        printf("%u: ", current_vertex);
         for (uint64_t idx = 0; idx + g->IA[current_vertex] < g->IA[current_vertex + 1]; current_vertex++)
         {
-            printf(" %lu", g->JA[idx + g->IA[current_vertex]]);
+            printf(" %u", g->JA[idx + g->IA[current_vertex]]);
         }
         printf("\n");
         //     edge *p; /* temporary pointer */
